[Bug/fix 1.0.4] 버그 수정

+++ app/src/main/java/com/mbj/ssassamarket/ui/bindings/TextViewBinding.kt
@@ -48,7 +48,7 @@ fun TextView.setFormattedElapsedTime(createdDate: String) {
 }
 
 @BindingAdapter("priceFormatted")
-fun TextView.setFormattedPrice(price: Int) {
+fun TextView.setFormattedPrice(price: Long) {
     val context = this.context
     val decimalFormat = DecimalFormat(PATTERN_COMMA_THOUSANDS)
     val formattedPrice = decimalFormat.format(price)
